refactor: Price 값 long으로 수정

+++ src/main/java/cart/dto/ProductResponse.java
@@ -6,9 +6,9 @@ public class ProductResponse {
     private final Long id;
     private final String name;
     private final String image;
-    private final Integer price;
+    private final long price;
 
-    public ProductResponse(Long id, String name, String image, Integer price) {
+    public ProductResponse(Long id, String name, String image, long price) {
         this.id = id;
         this.name = name;
         this.image = image;
@@ -31,7 +31,7 @@ public class ProductResponse {
         return image;
     }
 
-    public Integer getPrice() {
+    public long getPrice() {
         return price;
     }
 }
